# openapi-relevant.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IoT Sensor Platform API",
    "version": "0.2.0",
    "description": "Combined current contract (P1 + implemented P2)",
    "x-contract-scope": "relevant-p1-p2"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local instance"
    }
  ],
  "security": [
    {
      "jwtBearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "jwtBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT access token (P2 user auth)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "HTTP-only refresh cookie"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Entity not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProvisionRequest": {
        "type": "object",
        "required": [
          "serial",
          "powerSource"
        ],
        "properties": {
          "serial": {
            "type": "string",
            "pattern": "^SENS-[A-Z]{1,2}-\\d{5}$"
          },
          "displayName": {
            "type": "string",
            "maxLength": 255
          },
          "powerSource": {
            "type": "string",
            "enum": [
              "battery",
              "wired"
            ]
          },
          "zoneId": {
            "type": "string",
            "format": "uuid"
          },
          "calibrationOffsetC": {
            "type": "number",
            "default": 0
          }
        }
      },
      "ProvisionData": {
        "type": "object",
        "required": [
          "serial",
          "deviceType",
          "mqtt"
        ],
        "properties": {
          "serial": {
            "type": "string"
          },
          "deviceType": {
            "type": "string",
            "enum": [
              "TH",
              "TP",
              "T",
              "HM"
            ]
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "mqtt": {
            "type": "object",
            "required": [
              "username",
              "password",
              "topic",
              "statusTopic"
            ],
            "properties": {
              "username": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "topic": {
                "type": "string"
              },
              "statusTopic": {
                "type": "string"
              }
            }
          }
        }
      },
      "ProvisionEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/ProvisionData"
          }
        }
      },
      "Device": {
        "type": "object",
        "required": [
          "serial",
          "deviceType",
          "powerSource",
          "connectivityStatus",
          "alertStatus",
          "provisionedAt"
        ],
        "properties": {
          "serial": {
            "type": "string"
          },
          "deviceType": {
            "type": "string",
            "enum": [
              "TH",
              "TP",
              "T",
              "HM"
            ]
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "zoneName": {
            "type": "string"
          },
          "locationName": {
            "type": "string"
          },
          "powerSource": {
            "type": "string",
            "enum": [
              "battery",
              "wired"
            ]
          },
          "connectivityStatus": {
            "type": "string",
            "enum": [
              "online",
              "offline"
            ]
          },
          "alertStatus": {
            "type": "string",
            "enum": [
              "normal",
              "alert"
            ]
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastTemperatureC": {
            "type": "number",
            "nullable": true
          },
          "lastHumidityPct": {
            "type": "number",
            "nullable": true
          },
          "lastBatteryPct": {
            "type": "integer",
            "nullable": true
          },
          "provisionedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeviceEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/Device"
          }
        }
      },
      "DeviceListEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Device"
            }
          }
        }
      },
      "PatchDevice": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 255
          },
          "zoneId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "calibrationOffsetC": {
            "type": "number"
          }
        }
      },
      "DecommissionData": {
        "type": "object",
        "required": [
          "serial",
          "decommissionedAt"
        ],
        "properties": {
          "serial": {
            "type": "string"
          },
          "decommissionedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DecommissionEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/DecommissionData"
          }
        }
      },
...
      "CreateAlertRule": {
        "type": "object",
        "required": [
          "metric",
          "operator",
          "threshold"
        ],
        "properties": {
          "metric": {
            "type": "string",
            "enum": [
              "temperature_c",
              "humidity_pct"
            ]
          },
          "operator": {
            "type": "string",
            "enum": [
              "gt",
              "lt",
              "gte",
              "lte"
            ]
          },
          "threshold": {
            "type": "number"
          },
          "cooldownMinutes": {
            "type": "integer",
            "minimum": 1,
            "default": 15
          }
        }
      },
      "PatchAlertRule": {
        "type": "object",
        "properties": {
          "threshold": {
            "type": "number"
          },
          "operator": {
            "type": "string",
            "enum": [
              "gt",
              "lt",
              "gte",
              "lte"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "cooldownMinutes": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "AlertRule": {
        "type": "object",
        "required": [
          "id",
          "deviceSerial",
          "metric",
          "operator",
          "threshold",
          "isActive",
          "cooldownMinutes",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deviceSerial": {
            "type": "string"
          },
          "metric": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "threshold": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "cooldownMinutes": {
            "type": "integer"
          },
          "lastTriggeredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AlertRuleEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/AlertRule"
          }
        }
      },
      "AlertRuleListEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertRule"
            }
          }
        }
      },
...
      "NullEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "null"
          }
        }
      },
...
      "GenericOk": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": [
              "object",
              "array",
              "string",
              "number",
              "boolean",
              "null"
            ]
          }
        }
      }
    }
  },
  "paths": {
...
    "/devices/provision": {
      "post": {
        "operationId": "provisionDevice",
        "summary": "Register new device and generate MQTT credentials",
        "parameters": [
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisionEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/devices": {
      "get": {
        "operationId": "listDevices",
        "summary": "List all active devices",
        "responses": {
          "200": {
            "description": "OK (empty result = 200 with data: [])",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/devices/{serial}": {
      "get": {
        "operationId": "getDevice",
        "summary": "Get single device",
        "parameters": [
          {
            "name": "serial",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "patch": {
        "operationId": "updateDevice",
        "summary": "Update device properties",
        "parameters": [
          {
            "name": "serial",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchDevice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "decommissionDevice",
        "summary": "Decommission device (soft delete)",
        "parameters": [
          {
            "name": "serial",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decommissioned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecommissionEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/devices/{serial}/alert-rules": {
      "post": {
        "operationId": "createAlertRule",
        "summary": "Create alert rule for device",
        "parameters": [
          {
            "name": "serial",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRule"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "operationId": "listAlertRules",
        "summary": "List alert rules for device",
        "parameters": [
          {
            "name": "serial",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK (empty result = 200 with data: [])",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleListEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/alert-rules/{id}": {
      "patch": {
        "operationId": "updateAlertRule",
        "summary": "Update alert rule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAlertRule"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "deleteAlertRule",
        "summary": "Delete alert rule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/devices/{serial}/rotate-mqtt": {
      "post": {
        "operationId": "rotateDeviceMqtt",
        "summary": "Rotate device MQTT credentials",
        "description": "Admin-only operation. Returns new MQTT credentials; plaintext password is returned once.",
        "parameters": [
          {
            "name": "serial",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericOk"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
...
}